feat(service-form): support service spending time
Added a nnew section "Beneficiary participation" with spending_time_* fields (total hours and operating details)

Needed to enrich / fix the basic-input-field component in order to correctly support numeric value.

In order to do this, we needed to add type `number` to variable `value`. Thus, we had to force `value.toString()` in some places in the component. Not very satisfying, but I did not have clever idea.