test: configure project to allow component testing
* Add testing-library framework for Svelte (with jsdom)
* Fix TypeScript config
* Add an example of component testing with BasicInputField

Good-to-konw:
In order to render, <BasicInputField/> requires that the given ID to be included (defined) in the associated form schema. It is why the tests use `spendingTimeTotalHours`.

This mechanism is based on a Svelte $store object (so the '$' character in the template). I thought I had to defined "context" (in `render()` method) but they are two separate concepts. The solution was simplest, just importing and changing value of the form schema.