refactor: add services.utils#update_geom(service)
This method updates a service geom by calling BAN (Base Addresse Nationale) API, even if it already has a geometry property.

If the scoring of the result(s) is low (i.e. under 0.5/1) nothing is done.

This method is extracted from Django command "fix_missing_geom" in order to be reused in other import scripts.

The "assert" at the beginning of the command was not taken, in order to make this method more flexible in other context.